fix: Skip parent types if child matches too
As shown in #137, if a DOCX file stream is validated, it matches both DOCX and ZIP (which it depends on), resulting in `TryFindUnambiguousAsync` being kinda wrong.
In almost every case, we want the DOCX type and not the ZIP type.
Therefore, I've added `FindCloseMatchesAsync`, which filters found parent types.
`TryFindUnambiguousAsync` will make use of this. (Breaking change!)
For this to work every time, I had to make `FileByteFilter` abstract so that you can't work with "anonymous" file types (which would always be filtered out by any other file type class). (Breaking change!)
I also had to update almost all the test cases.